requirements.txt
PySide6>=6.0.0
gevent>=24.2.1
ffmpeg-python>=0.2.0